+++ src/Sonic/SRS.hs
@@ -3,9 +3,10 @@
 module Sonic.SRS where
 
 import Protolude
-import Sonic.Curve (Fr, G1, G2, GT, gG1, gG2)
 import Pairing.Pairing (reducedPairing)
-import Curve (Curve(..))
+import Curve (Curve(..), Group(..))
+
+import Sonic.Curve (Fr, G1, G2, GT)
 
 data SRS = SRS
   { srsD :: Int
@@ -20,20 +21,22 @@ data SRS = SRS
   , srsPairing :: GT
   }
 
+-- | Create a new Structured Reference String (SRS)
+-- 'd' should be large enough to support the circuit depth 'n'.
+-- In this implementation, 'd' should be greater than (7 * 'n')
 new :: Int -> Fr -> Fr -> SRS
 new d x alpha
   = let xInv = recip x
     in SRS
         { srsD = d
-        , gNegativeX = (mul gG1 . (^) xInv) <$> [1..d]
-        , gPositiveX = (mul gG1 . (^) x) <$> [0..d]
-        , hNegativeX = (mul gG2 . (^) xInv) <$> [1..d]
-        , hPositiveX = (mul gG2 . (^) x) <$> [0..d]
-        , gNegativeAlphaX = (mul gG1 . ((*) alpha . (^) xInv)) <$> [1..d]
+        , gNegativeX = mul gen . (^) xInv <$> [1..d]
+        , gPositiveX = mul gen . (^) x <$> [0..d]
+        , hNegativeX = mul gen . (^) xInv <$> [1..d]
+        , hPositiveX = mul gen . (^) x <$> [0..d]
+        , gNegativeAlphaX = mul gen . ((*) alpha . (^) xInv) <$> [1..d]
         -- g^alpha is not shared
-        , gPositiveAlphaX = (mul gG1 0) : (mul gG1 . ((*) alpha . (^) x) <$> [1..d])
-        , hNegativeAlphaX = (mul gG2 . ((*) alpha . (^) xInv)) <$> [1..d]
-        , hPositiveAlphaX = (mul gG2 . ((*) alpha . (^) x)) <$> [0..d]
-        , srsPairing = reducedPairing gG1 (mul gG2 alpha)
+        , gPositiveAlphaX = mul gen 0 : (mul gen . ((*) alpha . (^) x) <$> [1..d])
+        , hNegativeAlphaX = mul gen . ((*) alpha . (^) xInv) <$> [1..d]
+        , hPositiveAlphaX = mul gen . ((*) alpha . (^) x) <$> [0..d]
+        , srsPairing = reducedPairing gen (mul gen alpha)
         }
-